Test uint8

+++ components/tc_bus/tc_bus.cpp
@@ -152,10 +152,10 @@ namespace esphome
 
             #ifdef USE_API
 
-            auto api_userservicetrigger = new api::UserServiceTrigger<int32_t>("on_acknowledge", {"type"});
+            auto api_userservicetrigger = new api::UserServiceTrigger<uint8_t>("on_acknowledge", {"type"});
             api::global_api_server->register_user_service(api_userservicetrigger);
-            auto automation = new Automation<int32_t>(api_userservicetrigger);
-            auto lambdaaction = new LambdaAction<int32_t>([this](int32_t type) -> void {
+            auto automation = new Automation<uint8_t>(api_userservicetrigger);
+            auto lambdaaction = new LambdaAction<uint8_t>([this](uint8_t type) -> void {
                 TCBusComponent::on_acknowledge(12);
             });
             automation->add_actions({lambdaaction});
